Opuscodec: FIX memleak on decoding error
* When decoding fails, the codec is returning an empty chunk. At this
  time the original data must be freed.
* This change also avoids a memcpy for the decoded data when decoding
  was successful.

The result of realloc is discarded because: "If realloc() fails, the
original block is left untouched". The data is freed after playback.
This should never happen because realloc is always reducing the size of
the allocated data, however if it does - the result ist slightly more
memory usage for opus chunk sizes < 120ms.