client: avoid logfile-spam on buffer underrun
When audio buffer is empty print an error message should be printed for the buffer
underruns as those are audible and compensated with generated silence. Consecutive attempts to obtain an audio chunk from empty
buffer will lead to verbose messages instead of error - the effect is
already visible and audible but there is no new effect.

This closes #43