stream.c: avoid double add_fd()
inputpipe_resume_read() adds the inputpipe fd to epoll. When the stream
state is THROTTLE, this will be executed twice, causing a crash of the
application. This commits avoids that crash by dropping potentially
existing resume-tasks for this inputstream when running resume_read upon
first client appearing on the stream.